requirements.txt
fastapi
uvicorn
langchain
faiss-cpu
openai
streamlit
requests
mlflow
PyPDF2
numpy
pytest
langchain_huggingface